Move StoreFiles to backup.rs
